fix(trusty-mpm): no activity bump on Interrupt + align observe usize signature + approx-float test (review follow-up, refs #1461)
- inject(): only bump last_activity_at for Submit::Enter/NoSubmit; an
  Interrupt (C-c) is a STOP signal, not forward progress, so bumping it
  would mislead the idle/orphan-GC reconciliation. Add a dedicated test
  asserting Interrupt leaves last_activity_at None while Enter/NoSubmit
  set it.
- Align the observe line-count to usize end-to-end: ManagedTmuxDriver::
  capture, SessionManager::observe, and capture_pane now take usize,
  removing the load-bearing u32::try_from narrowing in
  DaemonSessionControl::observe. The single u32 conversion the tmux -S
  argv requires is confined to the RealTmuxDriver edge.
- verdict_to_summary_keeps_real_summary: use an approx float comparison
  (abs diff < 1e-6) instead of assert_eq! on an f64.

+++ crates/trusty-mpm/src/runtime/test_helpers.rs
@@ -61,7 +61,7 @@ impl ManagedTmuxDriver for FakeTmux {
         Ok(())
     }
 
-    fn capture(&self, _name: &str, _lines: u32) -> Result<String, ManagedError> {
+    fn capture(&self, _name: &str, _lines: usize) -> Result<String, ManagedError> {
         Ok(String::new())
     }
